+++ src/C3.Domain/Profiles/RepresentationPurpose.cs
@@ -1,0 +1,8 @@
+namespace C3.Domain.Profiles
+{
+    public enum RepresentationPurpose
+    {
+        DirectSave = 0,
+        ExportCopy = 1
+    }
+}